+++ src/rics/__init__.py
@@ -18,6 +18,6 @@ __all__ = [
     "configure_stuff",
 ]
 
-__version__ = "6.0.4.dev1"
+__version__ = "6.1.0"
 
 _logging.getLogger(__name__).addHandler(_logging.NullHandler())