+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.1.0] - 2025-12-08
+
 ### Added
 - New class `strings.ReprFormatter`.
 - Function `strings.format_kwargs()`, backed by `strings.ReprFormatter`.
@@ -592,7 +594,8 @@ First release on PyPI!
 - The `rics.mapping` package for linking elements in multiple directions.
 - The `rics.cardinality` package; enum types for `1:1`, `1:N`, `N:1`, and `M:N`.
 
-[Unreleased]: https://github.com/rsundqvist/rics/compare/v6.0.4...HEAD
+[Unreleased]: https://github.com/rsundqvist/rics/compare/v6.1.0...HEAD
+[6.1.0]: https://github.com/rsundqvist/rics/compare/v6.0.4...v6.1.0
 [6.0.4]: https://github.com/rsundqvist/rics/compare/v6.0.3...v6.0.4
 [6.0.3]: https://github.com/rsundqvist/rics/compare/v6.0.2...v6.0.3
 [6.0.2]: https://github.com/rsundqvist/rics/compare/v6.0.1...v6.0.2